+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 * Fix `Blueprint.new` handling of caching when the wrapped object responds to `identity_map`, but does not have one set.
+* Undefine JRuby package helper methods in `Model` (org, java...)
 
 ## 1.3.1